+++ covalent/_shared_files/defaults.py
@@ -27,6 +27,8 @@ from typing import Dict, List
 
 import dask.system
 
+from .config import CMType
+
 prefix_separator = ":"
 
 parameter_prefix = f"{prefix_separator}parameter{prefix_separator}"
@@ -62,6 +64,51 @@ def get_default_sdk_config():
             + "/covalent/executor_plugins"
         ),
         "no_cluster": "false",
+    }
+
+
+def get_default_server_config():
+    """
+    Return the default server config for a client
+    """
+    return {
+        "address": os.environ.get("COVALENT_DISPATCHER_ADDR")
+        if os.environ.get("COVALENT_DISPATCHER_ADDR")
+        else "localhost",
+        "port": int(os.environ.get("COVALENT_SVC_PORT"))
+        if os.environ.get("COVALENT_SVC_PORT")
+        else 48008,
+    }
+
+
+def get_default_service_config():
+    """
+    Return the default service configuration for the server (a.k.a dispatcher)
+    """
+    return {
+        "address": "0.0.0.0" if os.environ.get("COVALENT_SERVER_IFACE_ANY") else "localhost",
+        "port": int(os.environ.get("COVALENT_SVC_PORT"))
+        if os.environ.get("COVALENT_SVC_PORT")
+        else 48008,
+        "dev_port": int(os.environ.get("COVALENT_DEV_PORT"))
+        if os.environ.get("COVALENT_DEV_PORT")
+        else 49009,
+        "results_dir": os.environ.get("COVALENT_RESULTS_DIR", "results"),
+        "cache_dir": os.environ.get("COVALENT_CACHE_DIR")
+        or (
+            (os.environ.get("XDG_CACHE_HOME") or (os.environ.get("HOME") + "/.cache"))
+            + "/covalent"
+        ),
+        "log_dir": os.environ.get("COVALENT_LOGDIR")
+        or (
+            (os.environ.get("XDG_CACHE_HOME") or (os.environ.get("HOME") + "/.cache"))
+            + "/covalent"
+        ),
+        "db_path": os.environ.get("COVALENT_DATABASE")
+        or (
+            (os.environ.get("XDG_DATA_HOME") or (os.environ.get("HOME") + "/.local/share"))
+            + "/covalent/dispatcher_db.sqlite"
+        ),
     }
 
 
@@ -101,7 +148,7 @@ def get_default_workflow_data_config():
         "storage_type": "local",
         "base_dir": os.environ.get("COVALENT_DATA_DIR")
         or (
-            (os.environ.get("XDG_DATA_HOME") or (os.environ["HOME"] + "/.local/share"))
+            (os.environ.get("XDG_DATA_HOME") or (os.environ.get("HOME") + "/.local/share"))
             + "/covalent/workflow_data"
         ),
     }
@@ -128,17 +175,31 @@ def get_default_executor() -> dict:
     """
     from .config import get_config
 
-    return "local" if get_config("sdk.no_cluster") else "dask"
+    return "local" if get_config(CMType.CLIENT, "sdk.no_cluster") else "dask"
+
+
+def get_default_config_path(type: CMType) -> str:
+    base = (
+        os.environ.get("COVALENT_CONFIG_DIR")
+        or os.environ.get("XDG_CACHE_HOME")
+        or os.environ["HOME"] + "/.config"
+    )
+    filename = "/covalent.conf" if type == CMType.CLIENT else "/covalentd.conf"
+    return base + filename
 
 
 # Default configuration settings
 @dataclass
-class DefaultConfig:
+class DefaultClientConfig:
     sdk: Dict = field(default_factory=get_default_sdk_config)
-    dispatcher: Dict = field(default_factory=get_default_dispatcher_config)
+    server: Dict = field(default_factory=get_default_server_config)
+
+
+@dataclass
+class DefaultServerConfig:
+    service: Dict = field(default_factory=get_default_service_config)
     dask: Dict = field(default_factory=get_default_dask_config)
     workflow_data: Dict = field(default_factory=get_default_workflow_data_config)
-    user_interface: Dict = field(default_factory=get_default_ui_config)
 
 
 @dataclass